requirements.txt
anthropic>=0.40.0
apify-client>=1.8.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
